+++ domain/src/main/java/com/cirodev/boldweatherapi/domain/model/Current.kt
@@ -2,12 +2,12 @@ package com.cirodev.boldweatherapi.domain.model
 
 data class Current(
     val lastUpdated: String,
-    val tempC: Int,
+    val tempC: Double,
     val condition: Condition,
     val windKph: Double,
     val windDir: String,
     val precipMm: Double,
-    val humidity: Int,
-    val feelsLikeC: Int,
-    val uv: Int
+    val humidity: Double,
+    val feelsLikeC: Double,
+    val uv: Double
 )
